Begun Research on new (temporary) map
Here's what research have found during scene preparation:
- In 'UnitClick' script there was part of code still referring to  'ItemList' which generated some errors when LMB was clicked
- 'UnitControlScript' is for some reason hidden in 'selection circle'

I changed the following things:
- removed commented out parts of code in 'BuildingScript' and changed its name to 'ItemListScript'
- Moved 'selection circle' out of Units models's child to the same level in hierarchy as them.

Tasks to do for a next couple of working sessions will be added on Discord Server shortly after this commit
