+++ RTS Multiplayer/Assets/Scenes/Research Scene/Research Scene/NavMesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 816a5eb70e450a44bac5bdeb993f926d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
