+++ RTS Multiplayer/Assets/Temporary.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db89811af0e5e3049800aecf48ecf4f4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
