+++ RTS Multiplayer/Assets/Temporary/Materials.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 3ec29d8a24cfa3147b06fa0a43717959
-NativeFormatImporter:
+guid: 51f77f60dbdbbee4283137993878fd89
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 36300000
   userData: 
   assetBundleName: 
   assetBundleVariant:
