+++ RTS Multiplayer/Assets/Temporary/Models FBX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 243b241ad74ad5049a7a62c61dcc880f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
