+++ RTS Multiplayer/Assets/Temporary/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6dcfdeae1217ba44d9f48266c3052107
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
